backend/requirements.txt
arrow==0.12.1
asn1crypto==0.24.0
binaryornot==0.4.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
Django==2.0.6
django-cors-headers==2.3.0
django-filter==1.1.0
django-guardian==1.4.9
djangorestframework==3.8.2
future==0.16.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
jinja2-time==0.2.0
Markdown==2.6.11
MarkupSafe==1.0
mysql-connector==2.1.6
mysql-connector-python==8.0.11
mysqlclient==1.3.13
poyo==0.4.1
protobuf==3.6.0
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.5
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
whichcraft==0.4.1
wincertstore==0.2
